Add github actions workflow to run tests
Note that it seems like tests don't pass right now, I'm unsure if this is an issue with the test configuration in GitHub Actions or if it's an issue with some regressions on the tests.
Also note that this uses the Selenium Docker image in the steps of the workflow instead of as a service container, when I tried to use it as a service container it had issues that I believe were rooted in the fact that I couldn't attach it to the host network.