table with 3 plots per row or user defined layout
if the user has given input of row and column in theme animint, plots will be put in that location. but user needs to mention rows and columns for all plots. if the user does not want that option older functionality will work but with a new change of 3 plots per row and then a new row.