theme_animint now accepts dimensions and not theme
changed from theme to theme_animint for taking row and col arguments, removed the code to show browser side error if the user fails to add only row or col. because that is not necessary now because code to show error message in R session itself is there now. also changed theme function to theme_animint in the test i had wrote for passing row and col arguments